requirements.txt
streamlit
easyocr
opencv-python
numpy
Pillow
rapidfuzz
jiwer